change: business rules of modules

+++ backend/src/accounts/accounts.service.ts
@@ -3,11 +3,16 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Model } from 'mongoose';
 import { Account } from './schemas/account.schema';
 import { CreateAccountDto, UpdateAccountDto } from './dto/accounts.dto';
+import { Transaction } from 'src/transactions/schemas/transaction.schema';
+import { Transfer } from 'src/transfers/schemas/transfer.schema';
 
 @Injectable()
 export class AccountsService {
   constructor(
     @InjectModel(Account.name) private readonly Account: Model<Account>,
+    @InjectModel(Transaction.name)
+    private readonly Transaction: Model<Transaction>,
+    @InjectModel(Transfer.name) private readonly Transfer: Model<Transfer>,
   ) {}
 
   async showAccounts(userId: string) {
@@ -33,6 +38,20 @@ export class AccountsService {
   }
 
   async deleteAccount(accountId: string) {
+    const transactionsUsedAccount = await this.Transaction.find({
+      account: accountId,
+    });
+
+    if (transactionsUsedAccount.length > 0)
+      throw new Error('There are transactions using this account');
+
+    const transfersUsedAccount = await this.Transfer.find({
+      $or: [{ fromAccount: accountId }, { toAccount: accountId }],
+    });
+
+    if (transfersUsedAccount.length > 0)
+      throw new Error('There are transfers using this account');
+
     await this.Account.findByIdAndDelete(accountId).orFail();
     return { accountId };
   }
